+++ src/numerics/imag_axes_ft/IAFT.hpp
@@ -68,12 +68,12 @@ namespace imag_axes_ft {
         iaft_pt_exists = true;
         auto iaft_pt = child_pt.get();
         auto eps = read_eps_option(iaft_pt, "eps");
-        auto prec = read_prec_option(iaft_pt, "prec", eps.has_value() ? std::nullopt : std::optional<std::string>{"high"});
+        auto prec = read_prec_option(iaft_pt, "prec", eps.has_value() ? std::nullopt : std::optional<std::string>{"medium"});
 
         init_grid_variant(
           io::get_value_with_default<double>(pt,"beta",1000.0), 
           io::get_value_with_default<double>(iaft_pt,"wmax",wmax_default), 
-          string_to_basis_enum(io::get_value_with_default<std::string>(iaft_pt, "basis", "ir")), 
+          string_to_basis_enum(io::get_value_with_default<std::string>(iaft_pt, "basis", "dlr")), 
           prec,
           eps,
           print_meta_log
@@ -83,12 +83,12 @@ namespace imag_axes_ft {
       if (not iaft_pt_exists) {
         // Fall back to the old interface if "iaft" child ptree does not exist.
         auto eps = read_eps_option(pt, "iaft_eps");
-        auto prec = read_prec_option(pt, "iaft_prec", eps.has_value() ? std::nullopt : std::optional<std::string>{"high"});
+        auto prec = read_prec_option(pt, "iaft_prec", eps.has_value() ? std::nullopt : std::optional<std::string>{"medium"});
     
         init_grid_variant(
           io::get_value_with_default<double>(pt,"beta",1000.0), 
           io::get_value_with_default<double>(pt,"iaft_wmax",wmax_default), 
-          string_to_basis_enum(io::get_value_with_default<std::string>(pt, "iaft_basis", "ir")), 
+          string_to_basis_enum(io::get_value_with_default<std::string>(pt, "iaft_basis", "dlr")), 
           prec,
           eps,
           print_meta_log
@@ -96,7 +96,7 @@ namespace imag_axes_ft {
       }
     } 
 
-    IAFT(double beta, double wmax, basis_e basis, std::string prec="high", bool print_meta_log = false) {
+    IAFT(double beta, double wmax, basis_e basis, std::string prec="medium", bool print_meta_log = false) {
       init_grid_variant(beta, wmax, basis, prec, std::nullopt, print_meta_log);
     }